moved computation related to triangular grid to triangle_mod and fixed a nasty oversight.
okay so there are ICON grids that are really weirdly shaped, and one vertex is approx. 240 degrees away from its center on a lat-lon grid. How? I do not know. But I assume that that is correct. Then the width method earlier no longer works and I have to define a box with the max and min vertices of the triangle. Hopefully this catches all weird triangle cases.