caught two bugs in topo_mod; more details in longer commit msg
so I did not initialise cond_lat and cond_lon, and so the logical .and. check gave undefined behaviour. second bug is the duplicate checker for the lat_ and lon_indices arrays assumed that duplicated arrays are always stored on neighbouring indices. This is not true for all USGS topographical grid cells.